+++ src/asp/Camera/CsmModel.h
@@ -38,8 +38,8 @@ namespace csm {
 
 namespace asp {
 
-  // Do not set this lower than 1e-8, as then UsgsAstroLsSensorModel will
-  // return junk.
+  // Do not set this lower than 1e-8, as then UsgsAstroLsSensorModel can return
+  // junk because of numerical precision issues for high focal length.
   const double DEFAULT_CSM_DESIRED_PRECISION = 1.0e-8;
   
   /// Class to load any cameras described by the Community Sensor Model (CSM)